Split MaintenanceController into multiple Controller classes
